move:Buttonコンポーネントのファイル構成を修正
